feat(#S11P31D102-111 ) : 데이터 넘겨주기

+++ android/RunIt/watch/src/main/java/com/zoku/watch/model/ExerciseResult.kt
@@ -2,12 +2,13 @@ package com.zoku.watch.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import java.time.Duration
 
 
 @Parcelize
 data class ExerciseResult(
     val distance: Double?,
     val pace: Double?,
-    val time: Double?,
+    val time: Long?,
     val bpm: Double?
 ) : Parcelable
